Prevent call to the login method for non cookie based authentication
It seems that making a call to the login method in the constructor fail
in some cases even this should not happened (as ids are passed in the
URI by default, login should always return true).

The login is not necessary for non cookie based authentication but is
required for cookie based authentication. Also, it is preferable to
ensure the connection work in the constructor (instead of throwing
an exception at the first call).
For cookie based authentication, the login is still done but for simple
authentication this has been replaced by a call retrieving the version.
For having coherent message between simple and cookie authentication,
the call to the version catch 401 errors to show the same message when
ids are not valid.

This should fix #15 and #16